Project/production/python/requirements.txt
mysql-connector
jinja2
sanic
sanic_auth